refactor: ErrorCode로 네이밍 수정
