+++ src/main/java/roomescape/date/domain/ReservationDate.java
@@ -7,7 +7,7 @@ import roomescape.date.exception.ReservationDateException;
 
 import java.time.LocalDate;
 
-import static roomescape.date.exception.ReservationDateExceptionInformation.*;
+import static roomescape.date.exception.ReservationDateErrorCode.*;
 
 @Getter
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
