+++ src/main/java/roomescape/date/exception/ReservationDateException.java
@@ -1,10 +1,10 @@
 package roomescape.date.exception;
 
-import roomescape.common.exception.ExceptionInformation;
+import roomescape.common.exception.ErrorCode;
 import roomescape.common.exception.RoomEscapeException;
 
 public class ReservationDateException extends RoomEscapeException {
-    public ReservationDateException(ExceptionInformation exceptionInformation) {
-        super(exceptionInformation);
+    public ReservationDateException(ErrorCode errorCode) {
+        super(errorCode);
     }
 }
