+++ src/main/java/roomescape/reservation/exception/ReservaitonErrorCode.java
@@ -3,11 +3,11 @@ package roomescape.reservation.exception;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
-import roomescape.common.exception.ExceptionInformation;
+import roomescape.common.exception.ErrorCode;
 
 @Getter
 @AllArgsConstructor
-public enum ReservaitonExceptionInformation implements ExceptionInformation {
+public enum ReservaitonErrorCode implements ErrorCode {
 
     RESERVATION_ID_IS_NULL(HttpStatus.BAD_REQUEST, "예약 ID는 필수입니다."),
     RESERVATION_NAME_IS_NULL(HttpStatus.BAD_REQUEST, "예약자 이름은 필수입니다."),
