+++ src/main/java/roomescape/reservation/exception/ReservationException.java
@@ -1,10 +1,10 @@
 package roomescape.reservation.exception;
 
-import roomescape.common.exception.ExceptionInformation;
+import roomescape.common.exception.ErrorCode;
 import roomescape.common.exception.RoomEscapeException;
 
 public class ReservationException extends RoomEscapeException {
-    public ReservationException(ExceptionInformation exceptionInformation) {
-        super(exceptionInformation);
+    public ReservationException(ErrorCode errorCode) {
+        super(errorCode);
     }
 }
