+++ src/main/java/roomescape/theme/exception/ThemeException.java
@@ -1,10 +1,10 @@
 package roomescape.theme.exception;
 
-import roomescape.common.exception.ExceptionInformation;
+import roomescape.common.exception.ErrorCode;
 import roomescape.common.exception.RoomEscapeException;
 
 public class ThemeException extends RoomEscapeException {
-    public ThemeException(ExceptionInformation exceptionInformation) {
-        super(exceptionInformation);
+    public ThemeException(ErrorCode errorCode) {
+        super(errorCode);
     }
 }
