+++ src/main/java/roomescape/time/exception/ReservationTimeException.java
@@ -1,10 +1,10 @@
 package roomescape.time.exception;
 
-import roomescape.common.exception.ExceptionInformation;
+import roomescape.common.exception.ErrorCode;
 import roomescape.common.exception.RoomEscapeException;
 
 public class ReservationTimeException extends RoomEscapeException {
-    public ReservationTimeException(ExceptionInformation exceptionInformation) {
-        super(exceptionInformation);
+    public ReservationTimeException(ErrorCode errorCode) {
+        super(errorCode);
     }
 }
