+++ src/test/java/roomescape/reservation/domain/ReservationTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static roomescape.reservation.domain.ReservationStatus.CANCELED;
 import static roomescape.reservation.domain.ReservationStatus.RESERVED;
-import static roomescape.reservation.exception.ReservaitonExceptionInformation.*;
+import static roomescape.reservation.exception.ReservaitonErrorCode.*;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
