+++ src/test/java/roomescape/theme/domain/ThemeTest.java
@@ -2,7 +2,7 @@ package roomescape.theme.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static roomescape.theme.exception.ThemeExceptionInformation.*;
+import static roomescape.theme.exception.ThemeErrorCode.*;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
